separate Loom and non-loom tests
